refactor(fc-units-override): split detection and promotion into separate concerns
The detection concern had grown to include the promotion helper that
discards override rows and merges their units into the
`fixed_charges` params handed to `Plans::OverrideService`. The name
no longer matched the responsibility, and mixing a pure params-shape
predicate set with a DB-mutating helper blurred the concern's
boundary.

This commit moves the promotion helper into its own
`FixedChargeUnitsOverridePromotionConcern`. The detection concern is
back to just the two params-shape predicates and their normalisation
helper. `UpdateOrOverrideFixedChargeService` and `UpdateService`
include both concerns (they need both detection and promotion);
`CreateService` keeps only the detection concern (subscription
creation has no pre-existing override rows to promote).

The promotion concern still expects the host service to expose a
`subscription` reader.

Spec coverage for the promotion helper moves into the new concern's
spec file with the same cases (no overrides, captured-only, caller
wins on collisions, mixed sets, string-keyed input).
